CLOUDFRONT/Invalidate-S3-Objects/requirements.txt
arrow==1.2.3
boto3==1.26.87
botocore==1.29.87
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==1.0.1
make==0.1.6.post2
MarkupSafe==2.1.2
python-dateutil==2.8.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
python-dotenv==1.0.0